feat: refactor design-system folders
