+++ packages/app/__stories__/FormButton.stories.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { ComponentStory, ComponentMeta } from "@storybook/react"
 import { action } from "@storybook/addon-actions"
 
-import { FormButton } from "../design-system"
+import { FormButton } from "../src/design-system"
 
 export default {
   title: "Base/Form/FormButton",
