+++ packages/app/__stories__/FormGroup.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ComponentStory, ComponentMeta } from "@storybook/react"
 
-import { FormInput, FormGroup, FormLabel, FormGroupMessage } from "../design-system"
+import { FormInput, FormGroup, FormLabel, FormGroupMessage } from "../src/design-system"
 
 export default {
   title: "Base/Form/FormGroup",
