+++ packages/app/__stories__/FormLabel.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ComponentStory, ComponentMeta } from "@storybook/react"
 
-import { FormLabel } from "../design-system"
+import { FormLabel } from "../src/design-system"
 
 export default {
   title: "Base/Form/FormLabel",
